split array into several files, added documentation
